feat: relay support new chain(Polygon Scroll Linea Mantle Arbitrum Op…

+++ types/cross_chain.go
@@ -10,7 +10,7 @@ type CrossChainPackageType uint8
 
 type (
 	ChannelID uint8
-	ChainID   uint16
+	ChainID   uint32
 )
 
 const (
